requirements.txt
einops==0.8.0
matplotlib==3.8.4
numpy==2.1.1
pillow==10.4.0
tensorboard==2.18.0
timm==0.9.16
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.2
wandb==0.18.2
